Move the addressbook_test to test folder
