+++ ethers-addressbook/src/lib.rs
@@ -32,21 +32,3 @@ impl Contract {
 pub fn contract<S: Into<String>>(name: S) -> Option<Contract> {
     ADDRESSBOOK.get(&name.into()).cloned()
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_tokens() {
-        assert!(contract("dai").is_some());
-        assert!(contract("usdc").is_some());
-        assert!(contract("rand").is_none());
-    }
-
-    #[test]
-    fn test_addrs() {
-        assert!(contract("dai").unwrap().address(Chain::Mainnet).is_some());
-        assert!(contract("dai").unwrap().address(Chain::MoonbeamDev).is_none());
-    }
-}
